backends: simplify the `process` iterator
Summary
=======

Introduce a dedicated event that allows for pre-processing discovered
entities, no longer conflating this with the other events.

Details
=======

Pre-processing newly discovered entities (e.g., modifying the symbols,
registering them with the code generators) was so far done at the start
of processing an event.

There are multiple problems with this approach:
1. the `process` iterator cannot run code that must happen *after* pre-
   processing but before code generation
2. procedures discovered through scanning of constants were treated as
   first-seen in the module the constant is discovered from, while
   they're actually first-seen in the module the constant is defined in
3. the implementation is complex and hard to understand, requiring
   additional look-aside data structures (`EventContext`)

Number 1 is going to be a problem when moving to a data-oriented design
for the backend IRs.

To resolve the listed problems, a new event (`bekDiscovered`) is
introduced, providing a dedicated customization point for entity pre-
processing. The event is fired whenever new entities become available
(but, for implementation simplicity, sometimes also when nothing
changed).

This comes at the cost of the `process` iterator itself becoming larger
(after expansion and inlining) and that it contain more `yield`
statements.