cgen: remove dead code (#1051)
## Summary

Remove an optimization/semantic-change from `cgen` that's not used
anymore.

## Details

The `isInactiveDestructorCall`/`notYetAlive` procedures was responsible
for detecting whether a destructor call or reset operation could be
safely omitted for a local.

This is a workaround for destructor injection being limited to wrapping
whole scopes in `try`/`finally` statements, but it finally stopped
working with the introduction of the MIR (~11 months ago), as the 'def'
statements for all locals requiring destruction are placed before the
`try` statement responsible for their destruction.

Since the proper fix will be implemented in the destructor injection
pass rather than in the code generators, `isInactiveDestructorCall`,
`notYetAlive` are removed and their usage sites are updated. The now-
unused `compat.getRoot` procedure is removed too.