remove the `.linearScanEnd` pragma (#1053)
## Summary

Remove the `.linearScanEnd` pragma, as it complicates planned changes
to the C code generator.

## Details

When using the C code generator, the pragma forced all branches up to
and including the `of` branch it's part of to be translated into a
chain of `if` statements. It's an optimization hint that doesn't affect
effective program behaviour.

Since `mirgen` doesn't translate the pragma, it never reaches `cgen`,
effectively turning the pragma into a no-op. While this could be fixed,
it's existence would then complicate moving to a goto-based control-
flow representation for the `CgNode` IR, as there the body of a
branch is no longer directly associated with a case statement.