fix: compiler crash when `errorMax` != 1 (#1054)
## Summary

Fix the compiler crashing when compiling with an error limit beyond
one and there's an error somewhere in top-level code.

## Details

If the compiler doesn't abort on the first error (i.e., when
`errorMax` is `1`), an `nkError` node in top-level code reaches into
the `modulelowering` pass, which was so far treated as an internal
error.

An `nkError` node reaching into `modulelowering` is now ignored and the
pass is also disabled when an error occurred, as the code generation
phase is known to be disabled at that point.