disable the `tfragment_alloc` test (#1055)
## Summary

Disable the `tfragment_alloc` test due to it sometimes failing during
CI runs.

## Details

Investigation into the test failure showed that the test failure is not
because of memory exhaustion, but rather because of, in case of
failure, more pages than usual being requested from the OS.

In all observed CI failures, only the MacOS target was affected, so the
test is only disabled there.