cgir: use dedicated node kind per symbol kind
Summary
=======

Remove the `cnkSym` node kind and replace it with a dedicated kind for
each entity `cnkSym` previously represented. This brings the `CgNode`
IR closer to the MIR.

Details
=======

The `CgNode` IR groups symbols of constants, globals, fields, and
procedures under a single node kind, requiring dispatching based
on the `TSym.kind`. Which entity some logic expects is also not
expressible without a separate assertion.

In order to:
* reduce the amount of dispatching in the code generators
* bring the `CgNode` IR closer to the MIR
* prevent bugs caused by symbol kinds not being considered

the `CgNode` IR is changed to also use a dedicated node kind per
symbol.

Each usage of `cnkSym` is replaced with usage of the node kind(s)
that the respective code generator logic expects, which is context
dependent. A positive effect of this is that it is now clear which
entities are processed where.

Finally, a small cleanup is performed in `ccgstmts.genAsmOrEmitStmt`:
only the field symbol handling of the `cnkSym` branch is kept, for
all other symbol kinds, the else statement suffices.