fix(vmgen): taking the address of named constants (#1067)
## Summary

Support taking the address of named constants for code running in the
VM (affects both compile-time evaluation and the VM backend).

## Details

Global constants (i.e., `const`) that aren't inlined directly were
treated like non-constant globals, which is wrong, as global constants
are accessed through a different opcode.

**Changes:**
* implement proper code generator for `cnkConst` in `vmgen.genSym` and
  `vmgen.genSymAddr`
* remove the associated workaround in the MIR temporary elimination
  pass -- temporaries created from constant projections (e.g., `c.x.y`,
  where `c` is a `const`) are now also considered by the optimizer
* remove loading constants of primitive type as literals (i.e.,
  inlining constants). This already happens at an earlier stage
  (currently `sem`)

As a preparation for eventually being able to take the address of a
`const x = 1`, `vmgen` no longer throws an internal error when
encountering a constant with primitive type (integers, floats, string)
that wasn't inlined.