internal: support multiple lifetimes for entities (#1068)
## Summary

Update the `injectdestructors` pass to support an entity (global or
local) being defined multiple times within a MIR body. A workaround for
this previous shortcoming is now obsolete and thus removed.

## Details

### Background

No new global variables must be introduced after module lowering. For
this reason, `transf` doesn't introduce replicas for symbols of global
variables when inlining the same `for`-loop body multiple times
(because of more than one `yield` statement).

This, however, is a problem for the `injectdestructors` pass, as it, so
far, only supported an location existing once (i.e., there only being a
single 'def' for it). As a workaround, `transf` had to introduce
temporary replicas of the symbol, where are then replaced with the
originating-from symbol after the `injectdestructors` pass is over.

### The problem

The workaround is scattered across multiple layers and all new
processing that touches symbols of globals needs to account for the
fact that the symbol might not be the "correct" one.

### Solution

The `injectdestructors` pass is updated to support entity locations'
having multiple lifetimes within a body:
* associate an entity with a list of `EntityInfo` rather than only a
  single one (each one representing a single lifetime/scope)
* replace `getAliveGraph` with `findScope`, which looks up the entity's
  scope corresponding to the given data-flow instruction

A downside is that more allocations are performed due to the usage of
`seq`, but this is going to be resolved by injecting destructors
during AST -> MIR translation.

Creating a temporary copy of the symbol and later restoring it is
removed.