mir: unify defects and errors in the backend (#1071)
## Summary

Consider `--panics:on` during AST -> MIR translation and move all
decision making regarding whether a call potentially raises an
exception out of the code generators.

## Details

Calls within the MIR already encode whether or not they have an
exceptional exit (i.e., potentially raise an exception). However,
`--panics:on` was not considered during the AST -> MIR translation,
meaning that all non-magic were always considered as potentially
raising an exception, even if pancis were enabled and the procedure
raised no catchable errors.

This leaves considering `--panics:on` to the code generators, which is
a problem, as all behaviour-related decision-making needs to happen
prior code generation.

Therefore, `mirgen` now consider whether panics are enabled. This
also means that the `Defect` versus `CatchableError` distinction
no longer exists after the MIR phase.

### Architectural changes

* the `canRaiseDisp` procedure is renamed to `canRaise` and moved to
  the `ast_query` module (where the other `canRaise` procedure are)
* in addition, `canRaiseDisp` is adjusted to account for imported
  JavaScript procedures (identified by the `sfInfixCall` flag; compared
  to imported C functions, it's possible for them to raise exceptions)
* the pre-existing `canRaise` overload is un-exported
* `mirgen` now uses the new panics-considering `canRaise` overload for
  deciding whether a call can raise

### Dedicated node kind

The `mnkCheckedCall` node kind is introduced, replacing the `geRaises`
flag. This is progress towards:
* removing the `effects` field from `MirNode` (so that call nodes can
  store the number of sub-nodes instead)
* bringing the MIR and `CgNode` IR closer together

The `mnkCheckedCall` node kind is then translated to the
`cnkCheckedCall` node kind. For knowing whether the a call has an
exceptional exit, only the node kind needs to be inspected.

Finally, the all places that handle the `mnkCall`/`cnkCall` node kind
now also handle the `mnkCheckedCall`/`cnkCheckedCall` node kind. When
pretty-printing, a `mnkCheckedCall` is rendered like a `mnkCall`, but
with a trailing `(raises)` annotation.

### Lifetime hooks

In order to ensure that lifetime-hooks with exceptional exits work the
same as they did previously, `injectdestructors` emits potentially
raising calls according to the `canRaise` query.

This needs to be addressed eventually , either by inserting the calls
prior to AST -> MIR translation (unlikely) or by disallowing the hooks
from raising exceptions.

### Warnings

`cgen` now only consider whether the call is checked when reporting
an `ObservableStores` warning. This means that without `--panics:on`,
the warning is reported a lot more often now.

Since the direction is to remove the warning by not performing
observable stores in the first place, this regression is deemed okay.