mir: fix two `toOpenArray`-related bugs (#1074)
## Summary

Translate the `toOpenArray` magic to a dedicated built-in MIR operator
and fix two bugs related to the magic:
* regression from dc2ff8a4c49f3eb57ce11d84aeb5d33aa46b5f42:
  modifications through a `var openArray` not being visible on the
  source array when the source array is modified as part of the
  `toOpenArray` call's argument expressions
* arrays modified only through `var openArray` sometimes not being
  destroyed, leading to memory leaks (refer to the
  `tconsider_var_openarray` test)

## Details

Calls to the `toOpenArray` magic are represented in the MIR via generic
magic calls, but this doesn't properly capture their behaviour. The
core problem is that the generic magic call doesn't establish a
connection between the array operand and the resulting `openArray`.

Since, at the moment, data-flow analysis doesn't track the roots of
`openArray` values, this also means that creation of a `var openArray`
via `toOpenArray` doesn't count as a mutation of the source value.

In order to resolve these problems, `toOpenArray` is translated to a
dedicated MIR operation.

**Changes**:
* introduce a three-argument variant of the `mnkToSlice` operation (a
  dedicated node kind might be better in the future, but overloading
  the operator is slightly simpler for the moment)
* translate the `mSlice` magic to the three-argument `mnkToSlice`
  variant
* translate the array operand in a way such that the value's identity
  is kept (`genOperand`). This fixes array values possibly being copied
* introduce a dedicated `cnkToSlice` node corresponding to
  `mnkToSlice`. This brings the two IRs closer together
* all three code generators continue to treat the single-argument
  `ToSlice` variant as a no-op for now
* the three-argument variant is treated the same as the `mSlice` magic
  previously was