internal: improve VM's object serialization (#1073)
## Summary

* simplify the logic for initializing a location from an object
  construction expression
* support arbitrarily shaped, but valid, object construction
  expressions

## Details

Serialization of object construction expression went over each field in
the destination location and initialized it to the value provided by
the expression. The logic relied on:
* all active fields (given the discriminator value(s)) having an
  associated initializer (`nkExprColonExpr`) in the construction AST
* the `nkExprColonExpr` being ordered by their field's position

This is unnecessarily complex and restrictive. Iterating over the AST
nodes and loading the values suffices, and it also doesn't put the
aforementioned restrictions on the AST.