vmgen: implement basic RVO support
Much of the work is done upfront, in `putIntoLoc` and
`genAsgnToLocal`. Besides the naming being inconsistent with `cgen`,
this will also lead to somewhat convoluted handling for magics that
return non-register-fitting values.

Implementing assignments similar to how `cgen` does it would be
better. However, it would also mean making much larger changes to
`vmgen`, which doesn't make sense this close to the rewrite.