vmgen: fix crash with macro API magics
Three magics part of the macro API are overloaded by name, meaning that
`vmgen` needs access to the originally used symbol. For now, the
magics are added to the `MagicsToKeep` set, so that `mirgen` doesn't
throw the symbols away.

Unfortunately, this means that a function table entry is,
unnecessarily, created for them.