vmgen: use less temporary registers (#1079)
## Summary

Don't use an unnecessary temporary register for `if` and `case`
conditions.

## Details

For an `if` and `case` statements, `vmgen` emits the following:
```
# for an `if`:
FastAsgnComplex r2 r1 # r1 holds the condition
FJmp r2, ...

# for a `case`:
FastAsgnComplex r2 r1 # r1 holds the selector value
Branch r2, ...
```

This is because `withTemp` allocates a temporary register itself, and
loading a local variable - which the condition/selector currently
always is, due to the structure of the MIR - into a register requires a
register copy (`opcFastAsgnComplex`).

`withTemp` is replaced with `withDest`, which doesn't allocate a
destination register, but still takes care of freeing it. This
eliminates the unnecessary register plus register copy.