mirgen: handle statement magics
`genx` assumed that all magics expanded to some expression, but this is
no longer true now.

For the fix, the `MirBuilder.push` procedure needs to support both
empty fragments and querying the length. Empty fragments require the
use of `HOslice` (half-open slice) instead of `Slice`, due to the index
type being unsigned.