lower `inc` and `dec` into assignments (#1089)
## Summary

The `inc a, b`/`a += b` and `dec a, b`/`a -= b` statements are now
lowered into `a = a + b` and `a = a - b`.

This makes the behaviour of both more similar: `inc a` raised an
`OverflowDefect` when `a` is of `range` or `enum` type and the result
is out-of-range, whereas `a = a + 1` raised a `RangeDefect` -- a
`RangeDefect` is now raised by both forms.

The aforementioned range check is now also performed when using the JS
or VM backend.

## Details

* lowering the `mInc` and `mDec` magic is done in `mirgen` (AST -> MIR
  translation)
* either `mAddI`/`mSubI` and `mAddU`/`mSubU` are used for the lowering,
  depending on the destination's signed-ness
* for `tyRange` and `tyEnum` types, an additional range check for the
  result of the addition/subtraction in inserted (but only if range
  checks are enabled)
* the `void` call handling in `mirgen.genx` now supports
  `genCallOrMagic` not emitting an expression (`mInc`/`mDec` lowering
  triggers this case)
* `vmgen.Loc` and the routines working with it were only used for the
  `mInc`/`mDec` implementation. They're obsolete now and thus removed
* the `trangechecks` and `tdistinct_range` tests are enabled for all
  backends

### Misc

`mirgen` now needs to `mirconstr.push` to support empty fragments. A
`HOslice[NodeIndex]` is used instead of `Slice[NodeIndex]` now, making
it possible to represent zero-length slices starting at index 0.