mirgen: fix moving/assigning to object conversions
Object conversions are special projections, and the way `mirgen`
captured the origin (always via `Bind`) was a problem for `jsgen` (and
it's also generally incorrect).

The conversion are now implemented in a dedicated procedure, with an
alias support assignments being used when required. Collapsing multiple
nested object conversions into a single one now also works for down
conversions.