fix(cgen): C compiler error with closures (#1095)
## Summary

The C code generated when assigning a `.nimcall` procedure to a
closure relied on an implicit conversion being introduced by the C
compiler, which some C compilers complain about.

An explicit cast is added so that the code works everywhere.

## Details

An extra configuration file is needed for the test, as the default C
compiler arguments have to be destructively overridden -- the `-w`
option passed by default would disable all warnings otherwise.