cgen: refactor `genDollar` (#1098)
## Summary

Remove `genDollar` and replace calls to it with calls to
`unaryExpr`. The workarounds implemented by `genDollar` are obsolete
now, and without them, `genDollar` does the same as `unaryExpr`.

## Details

The `d.k == locNone` condition in `genDollar` is never true, since
all magic calls reaching the code generator appear directly in the
source expression slots of assignments.