sem: `semFor` no longer mutates input AST (#1091)
## Summary

Stop mutating the input AST when semantically analysing `for` loops in
`semFor` .

## Details

Previously,  `semFor`  would mutate the input AST parameter  `n` ,
resulting in partial contribution to call analysis corruption during
overload resolution. Now  `semFor`  does a cursory examination of the 
`for`  loop and dispatches to either  `semForVars`  or  `semForFields` 
based on whether it's a standard  `for`  over non-type values or over a
type's fields, respectively.

As part of this change and issue was discovered in  `semAfterMacroCall`
, which was not handling errors passed via  `tyError` , this has been
updated by having it use  `semTypeNode2` . This meant a minor change in
error message for meta-routines returning a type ( `typedesc` ), from 
`expression has no type`  to 
`type expected, but expression has no type` , which is somewhat more
informative; tests have been updated accordingly.

---------

Co-authored-by: zerbina <100542850+zerbina@users.noreply.github.com>