koch, niminst: move csources bundle tag location
Previously, we stored the csources bundle tag using the confusing name
"csources-release" and stored within the build/csources folder.

However, this caused some undesirable issues:

- If this file become stored within the csources git repository, it will
  disable all csources versioning logic in koch.py.
- The old name "csources-release" did not adequately describe what the
  file is for.

This commit move the bundle tag to `build/csources-bundled-version`,
which now accurately describe the intent of the file and move the
ownership of "bundled" state away from csources.