osproc: don't allocate memory during fork-exec (#1107)
## Summary

Previously,  `osproc`  would attempt to allocate memory during fork-exec
via the use of  `findExe`  to find programs in  `PATH` . This causes
memory allocation during this sensitive period.

This PR replaces the use of  `findExe` , which was employed due to 
`execvpe`  not being ubiquitous, with  `execvpe` , now that it is widely
available.

## Details

There are a couple problems with allocating memory during fork-exec:

- This is extremely unsafe as this code is also used in  `vfork()`  case
(although disabled by default, thankfully), of which the use of an
allocator would completely mess up the parent process.
- This is super slow, since  `fork()`  has to perform CoW on pages that
are touched by the allocation procedure.
- A deadlock can occur in rare cases due to interaction with the global
allocator lock. This problem was observed in disruptek/balls.

With regard to substituting `findExe` with `execvpe`:

* This API has been available in uClibc since 2013, musl libc since 2014
and Haiku since 2015, removing the need to special case against those
systems.
* We can expect behaviors closer to how other programs perform  `PATH` 
searches. It is expected that there won't be any regression from this
change.