sem: ill-formed sym in `for` produces an error
Summary

Ill-formed `for` loops with an error in the sym position, in particular
produced by meta routines now correclty produce an error.

Details

Prior to this fix the error would not be reported because it was not
added to the production that was ultimately wrapped in an `nkError`.
This would result in a compiler crash due to assertion failure,
wrapping AST in an error without an error present therein.